+++ docs/source/changelog.rst
@@ -8,6 +8,7 @@ Change Log
 - test.echoserver now correctly deals with a specified hmac key in combination with name server usage
 - added connection troubleshooting checklist to tips & tricks chapter
 - fix sphinx config issues when building the docs
+- some raised exceptions had a __cause__ added on Python 2.x as well, this has been corrected (it could cause unwanted serialization errors)
 
 
 **Pyro 4.32**